Patches

+++ src/app/services/chapter.service.ts
@@ -25,23 +25,16 @@ export class ChapterService {
   ) {
     this.fs = (window as any).fs;
   }
-  // public bodyBlock: string;
-  // public notes: string;
-  // public notesArray: HTMLElement[] = [];
+
   public notes2: Note[] = [];
 
-  // public paragraphs: Paragraph[] = [];
-  // public header: string;
   public verseNums: number[] = [];
   public contextNums: number[] = [];
-  // public pageUrl = '';
+
   private fs: any;
-  // scripturesArchive: Observable<ArrayBuffer>;
-  // hebWTags: Document;
-  // wTagRefs: NodeListOf<Element>;
-  // verseSelect = false;
+
   chapter2: Chapter2 = new Chapter2();
-  // wTags: Array<[string, string, string]> = [];
+
   scrollIntoView: Element;
 
   public notes: ElementRef[] = [];
@@ -58,21 +51,17 @@ export class ChapterService {
     synchronizedScrolling: () => Promise<void>
   ): Promise<boolean> {
     return new Promise<boolean>((resolve, reject) => {
-      // this.paragraphs = [];
       this.notes2 = [];
       this.navService.pageTitle = '';
-      // this.header = '';
+
       let vSplit = chapter.split('.');
-      // this.hebWTags = new Document();
+
       if (chapter === '') {
         vSplit = book.split('.');
       }
 
       this.verseNums = this.parseHighlightedVerses2(vSplit[1]);
       this.contextNums = this.parseHighlightedVerses2(vSplit[2]);
-      // console.log(this.verseNums + ' ' + this.contextNums);
-
-      console.log(this.fs);
 
       if (this.fs) {
         this.getChapterFS(book, vSplit, synchronizedScrolling).then(
@@ -87,9 +76,6 @@ export class ChapterService {
           }
         );
       }
-      console.log('asodifjaoisdfjaoisdjfoiasdjfoiasdjfoij');
-
-      // resolve(true);
     });
   }
 
@@ -155,13 +141,12 @@ export class ChapterService {
             book.toLowerCase(),
             vSplit[0].toLowerCase()
           ) + '.json';
-        console.log(url2 + ' url2');
 
         this.fs.readFile(
           'c:/ScripturesProject/' + url2,
           'utf8',
           (err, data) => {
-            this.setChapter(data);
+            this.setChapter(data, synchronizedScrolling);
           }
         );
         resolve(true);
@@ -182,9 +167,7 @@ export class ChapterService {
         const saved = await localForage.getItem(book + '\\' + vSplit[0]);
 
         if (saved) {
-          console.log('asved');
-
-          this.setChapter(saved as string);
+          this.setChapter(saved as string, synchronizedScrolling);
         } else {
           const url2 =
             'assets/' + this.navService.urlBuilder(book, vSplit[0]) + '.json';
@@ -195,7 +178,7 @@ export class ChapterService {
               responseType: 'text'
             })
             .subscribe(data => {
-              this.setChapter(data);
+              this.setChapter(data, synchronizedScrolling);
               resolve(true);
             });
         }
@@ -203,30 +186,28 @@ export class ChapterService {
     );
   }
 
-  private async setChapter(data: string) {
+  private async setChapter(
+    data: string,
+    synchronizedScrolling: () => Promise<void>
+  ) {
     return new Promise<boolean>(
       async (
         resolve: (resolveValue: boolean) => void,
         reject: (rejectValue: boolean) => void
       ) => {
         this.chapter2 = JSON.parse(data) as Chapter2;
-        // this.resetVerseSelect();
+
         this.setHighlighting().then((value: boolean) => {
           this.verseFocus().then((value2: boolean) => {
+            // synchronizedScrolling();
             resolve(true);
           });
         });
-
-        console.log('verse focus');
-
-        // resolve(true);
       }
     );
   }
 
   public parseHighlightedVerses2(v: string): number[] {
-    // console.log('parseHighlightedVerses2');
-
     if (v === null || v === undefined) {
       return [];
     }
@@ -245,7 +226,6 @@ export class ChapterService {
 
     return verseNums;
   }
-  // public verses: ElementRef[] = [];
 
   public toggleVerseSelect() {
     this.saveState.data.verseSelect = !this.saveState.data.verseSelect;
@@ -271,48 +251,68 @@ export class ChapterService {
 
   public modifyWTag(callback: (elem: Element) => void) {
     this.wTags.forEach(e => {
-      console.log(e.nativeElement as [
-        string,
-        string,
-        string,
-        string,
-        string,
-        string,
-        string,
-        string,
-        string
-      ]);
-
       callback(e.nativeElement as Element);
     });
   }
 
-  public resetVerseSelect() {
-    // this.verseSelected = false;
-    console.log('resetVerseSelect');
-
+  public resetVerseNotes(
+    wTag: [
+      string,
+      string,
+      string,
+      string,
+      string,
+      string,
+      number,
+      string[]
+    ] = null
+  ) {
     this.resetNotes2();
+    this.resetVerseSelect(wTag);
+  }
 
+  public resetVerseSelect(
+    wTag: [
+      string,
+      string,
+      string,
+      string,
+      string,
+      string,
+      number,
+      string[]
+    ] = null
+  ) {
     this.modifyWTags(
       (
         wa: [string, string, string, string, string, string, number, string[]]
       ) => {
-        wa[7] = [];
+        if (wa !== wTag) {
+          wa[7] = [];
 
-        wa[0] = this.stringService.removeAttribute(wa[0], 'verse-select-0');
-        wa[0] = this.stringService.removeAttribute(wa[0], 'note-select-1');
-        if (this.saveState.data.verseSelect) {
-          this.createRefList(wa, this.saveState.data.newNotesVisible, 3);
-          this.createRefList(wa, this.saveState.data.englishNotesVisible, 4);
-          this.createRefList(wa, this.saveState.data.translatorNotesVisible, 5);
-          // console.log(wa[7]);
+          wa[0] = this.stringService.removeAttribute(wa[0], 'verse-select-0');
+          wa[0] = this.stringService.removeAttribute(wa[0], 'note-select-1');
+          if (this.saveState.data.verseSelect) {
+            this.createRefList(wa, this.saveState.data.newNotesVisible, 3);
+            this.createRefList(wa, this.saveState.data.englishNotesVisible, 4);
+            this.createRefList(
+              wa,
+              this.saveState.data.translatorNotesVisible,
+              5
+            );
 
-          if (wa[7].length !== 0) {
-            wa[0] = this.stringService.addAttribute(wa[0], 'verse-select-0');
-            wa[0] = this.stringService.removeAttribute(wa[0], 'verse-select-1');
+            if (wa[7].length !== 0) {
+              wa[0] = this.stringService.addAttribute(wa[0], 'verse-select-0');
+              wa[0] = this.stringService.removeAttribute(
+                wa[0],
+                'verse-select-1'
+              );
 
-            wa[0] = this.stringService.removeAttribute(wa[0], 'verse-select-2');
-            // console.log(wa[5]);
+              wa[0] = this.stringService.removeAttribute(
+                wa[0],
+                'verse-select-2'
+              );
+            }
           }
         }
       }
@@ -324,13 +324,9 @@ export class ChapterService {
     vis: boolean,
     noteNumber: number
   ) {
-    // console.log(vis);
-
     if (vis) {
       (wa[noteNumber] as string).split(',').forEach(w => {
         if (w.trim() !== '') {
-          // console.log(w);
-
           wa[7].push(w);
         }
       });
@@ -352,14 +348,9 @@ export class ChapterService {
   }
 
   public resetNotes2() {
-    console.log('asdoifjaosid fjaosidfjaoisdjfA');
-
     _.each<ElementRef>(this.notes, n => {
       (n.nativeElement as HTMLElement).classList.remove('verse-select-1');
     });
-    // _.each(this.chapterService.notes2, note => {
-    //   note.resetVerseSelect();
-    // });
   }
 
   public removeVerseSelect() {
@@ -375,18 +366,12 @@ export class ChapterService {
         }
       }
     );
-
-    // _.each(this.chapterService.wTags, wTag => {
-    //   wTag[0] = wTag[0].replace(' verse-select-0', '');
-    // });
   }
 
   public wTagClick(
     w: [string, string, string, string, string, string, number, string[]],
     verse: Verse
   ) {
-    // console.log(this.saveState.data.verseSelect);
-
     if (
       w[7].length === 0 &&
       !this.stringService.hasAttribute(w[0], 'verse-select-2') &&
@@ -395,55 +380,22 @@ export class ChapterService {
       return;
     }
     if (this.saveState.data.verseSelect) {
-      // console.log('asodifj');
-
       if (this.stringService.hasAttribute(w[0], 'verse-select-0')) {
         this.firstClick(w, verse);
       } else if (this.stringService.hasAttribute(w[0], 'verse-select-1')) {
-        this.resetVerseSelect();
+        this.resetVerseNotes();
       } else if (this.stringService.hasAttribute(w[0], 'verse-select-2')) {
-        // console.log('asdfasdfasdfasdfasdf');
         this.selectNote(w);
       }
     }
   }
-  resetVerseSelect1(): void {
-    this.modifyWTags(
-      (
-        wa: [string, string, string, string, string, string, number, string[]]
-      ) => {
-        if (wa[3].trim() !== '') {
-          wa[0] = this.stringService.addAttribute(wa[0], 'verse-select-0');
-          wa[0] = this.stringService.removeAttribute(wa[0], 'verse-select-1');
-        }
-        if (wa[4].trim() !== '') {
-          wa[0] = this.stringService.addAttribute(wa[0], 'verse-og-select-0');
-          wa[0] = this.stringService.removeAttribute(
-            wa[0],
-            'verse-og-select-1'
-          );
-        }
-        if (wa[5].trim() !== '') {
-          wa[0] = this.stringService.addAttribute(wa[0], 'verse-tc-select-0');
-          wa[0] = this.stringService.removeAttribute(
-            wa[0],
-            'verse-tc-select-1'
-          );
-        }
-      }
-    );
-  }
-
   private firstClick(
     w: [string, string, string, string, string, string, number, string[]],
     verse: Verse
   ) {
     this.resetVerseSelect();
-    // this.verseSelected = true;
-    // console.log(w[7]);
-    verse.wTags2.forEach(wr => {
-      // console.log(wr);
 
+    verse.wTags2.forEach(wr => {
       w[7].forEach(ref => {
         if (wr[7].includes(ref) && wr[7].length >= 1) {
           wr[0] = this.stringService.removeAttribute(wr[0], 'verse-select-0');
@@ -454,36 +406,36 @@ export class ChapterService {
         }
       });
     });
-    this.selectNote(w);
+    this.selectNote(w, false);
   }
 
   private refCount(refs: string): number {
     return refs.split(',').length;
   }
   private selectNote(
-    wTag: [string, string, string, string, string, string, number, string[]]
+    wTag: [string, string, string, string, string, string, number, string[]],
+    resetVerse: boolean = true
   ) {
-    // console.log(wTag[7].length);
     this.resetNotes2();
     if (wTag[7].length === 0) {
-      // console.log(wTag[7]);
-
-      this.resetVerseSelect();
+      this.resetVerseNotes();
       return;
     }
     const note = _.find(this.notes, (n: ElementRef) => {
       return (n.nativeElement as HTMLElement).id === wTag[7][0];
     });
     wTag[7].shift();
-    console.log(note);
 
     if (note) {
       (note.nativeElement as HTMLElement).classList.add('verse-select-1');
       (note.nativeElement as HTMLElement).scrollIntoView({
         block: 'center'
       });
+      if (resetVerse) {
+        this.resetVerseSelect(wTag);
+      }
     } else {
-      this.resetVerseSelect();
+      this.resetVerseNotes();
     }
   }
 }
